api/fast-api-requirements.txt
fastapi
pydantic
uvicorn
mysql-connector-python
jinja2
# OAuth2 https://www.youtube.com/watch?v=5GxQ1rLTwaU
python-multipart
# generate JWT (JSON web token)
python-jose[cryptography]
# Password hashes - bcrypt is one of the algos
passlib[bcrypt]
